Known issues cleanup  (#161)
* removed known issues that no longer apply

* moved a known issue in to troubleshoot section

* Delete helm_issue.md

removing this file because it is not applicable to RHACM limited release. It was also confirmed by Nathan via slack that "Our helm stuff never deployed on a managed cluster and the installer doesn't use helm directly. The subscription operator does, but I've not seen that entry before"

* Update known_issues.md

* Update known_issues.md

few more changes

* Update known_issues.md

making a comment

* Update about/known_issues.md

Co-Authored-By: James M. Young <jayoung@redhat.com>

* Update known_issues.md

made an update to the suggestion

* Update known_issues.md

Co-authored-by: James M. Young <jayoung@redhat.com>