open-cluster-management/backlog#123 ready f… (#181)
* https://github.com/open-cluster-management/backlog/issues/123 ready for team to write

@cadawson and @dockerymick these are the notes and what I started for us, please edit keeping in mind the notes from the call with Michael. I just took some info and got us started.

He has asked that we bring this first What's new topic, which will be used in the field, to the arch call (I will add on Thursday's agenda) for review, so if we know our areas and content enough on Thursday to get feedback, that will be great. Prob need to make sure links and parent topic links work by then, as well. I will do the same for my areas. Thank you!

* notes added

* Update whats_new.md

Added the blurb for security and compliance

* Update whats_new.md

* open-cluster-management/backlog#152 Update CL inst

* open-cluster-management/backlog#123 push what's new changes for meeting

Co-authored-by: Mikela Dockery <mdockery@redhat.com>
Co-authored-by: Christopher Dawson <cdawson@redhat.com>