3.11 on hub
I think we need to make this clear across install doc what versions are supported. 

When I redid supported clouds a couple weeks ago, there was some stuff that had been removed; I added the supported hub and managed info back into the doc. Devs confirmed on the hub that 3.11 is not supported at that time. 

Please check on the rest of this OS table:

https://github.com/open-cluster-management/rhacm-docs/issues/353