No issue: build fix attempt
Looks like these files are causing some issues with the build.  My thought is that the script might not differentiate between links in a comment and links out of a comment.  The titles of these commented out links did not match the titles in the actual files. I am hoping that by making them match, the issues might be resolved. Let me know if you have questions.