Improve the IAM Policy Documentation (#2749)
* Remove the IamPolicy namespaceSelector field

This field is here for legacy reasons but it does not affect the
behavior of the IAM policy controller, as the controller only handles
cluster roles (i.e. not namespaced roles). Removing the documentation
for this field makes things less confusing to the user.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Document the IAMPolicy spec.clusterRole field

This was added in 2.3 in the following commit but was not documented:
https://github.com/open-cluster-management/iam-policy-controller/commit/23b2c7d876ec16bafcb6f65b4b80bd4eb925faa1

This also fixes a couple small grammatical errors.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>