Add GHA to test AsciiDoc builds on each PR
Add a GitHub action that builds each .adoc file on each PR, failing if
there are any errors WARN-level or above. All issues are shown in the
logs before failing.

Subsequent PRs could lower the threshold for failing to INFO and fix
additional issues.

This could eventually be extracted to a stand-alone GitHub action,
perhaps moved to a different repo, and shared with the rest of the
AsciiDoc ecosystem.

Shared automated testing infra makes the contribution process more
comfortable for developers and the code review process easier for
owners. It also guarantees consistent standards across the project.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>