Update managing_hosted_aws.adoc
The steps indicated From the previous section" Important" and on detail the usage of the console with the UI providing more than just information. In this case, the UI only provides information on how to create a cluster.  but it does not walk you through it, which is the reason why these steps have been removed. Additionally, the console on bare metal does which is why this information needs to be moved to another document, which Roke has mentioned. This may need to be a separate request. I will tag @swopebe  to confirm if this issue needs to be a seperate one.
Additionally(@swopebe ) Roke has mentioned adding a note in which states the information presented above. I have added a mock note, and will refer back to the style guide as I am not quite sure of both a) the reference of UI and b)the necessary adjustments to my mock note. Please review but I will also review this and suggest changes.

 @chenz4027 if you could go ahead and review the changes to this document as you see fit. Thank you all for your help on one of my very first issues here as an Intern. I look forward  to your responses.