Update special_considerations.md simplify asterisks
With internal doc being shared across the team, we want to ensure we 1. don't document what writers should just know (like what is code and what is not) and 2. update this topic based on where people are running into issues but in the simplest way possible.

Asterisks are either part of the code if not bulleted lists or bold tags, and should always be treated as such. I added a recent example that wasn't really clear before how to fix lists and how to avoid formatting issues in DNS examples.