Fix the operator policy documentation
The instructions were incorrect on how to install an operator using
OperatorPolicy since we never recommend deploying a policy template
directly. We must always use a Policy object to deliver it.

After correcting this, the documentation was pretty much the same as the
policy_operator.adoc file, so I decided to remove it.

This also fixes a syntax error in the "Operator policy YAML table" to
allow a list to be rendered correctly in the spec.subscription table
cell.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>