Update policy_generator.adoc updates
We need to initiate an update anyway, but reading this, I felt a little disoriented because we introduce many samples at once with a lot of stuff, but didn't say much about them. For screen readers, I like to point out things early in the intro.

Also, since we only have one call out, we can add it to the top.